♻️ merge auth.ts into server.ts
