+++ src/auth.test.ts
@@ -1,5 +1,4 @@
 import { afterAll, beforeAll, describe, expect, it } from "bun:test";
-import { create_auth_handlers } from "./auth";
 import { parse_config } from "./config";
 import { DEV_PATH_PREFIX, TEST_USER } from "./dev/oidc_provider";
 import { create_server } from "./server";
@@ -33,8 +32,7 @@ const has_visible_text = async (view: Bun.WebView, text: string) =>
 
 beforeAll(() => {
   const session_store = create_session_store(config.SESSION_SECRET);
-  const auth = create_auth_handlers(config, session_store);
-  server = create_server(config.PORT, session_store, auth);
+  server = create_server(config.PORT, session_store, config);
 });
 
 afterAll(() => {
